refactor: dispatcher 클래스 삭제 및 디스패치 메서드 WAS로 이동
