+++ tomcat/src/main/java/org/apache/catalina/startup/Controllers.java
@@ -1,0 +1,16 @@
+package org.apache.catalina.startup;
+
+import com.techcourse.controller.Controller;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public class Controllers {
+    private final List<Controller> value;
+
+    public Controllers(Controller... controller) {
+        value = Arrays.stream(controller).toList();
+    }
+
+
+}
