+++ tomcat/src/main/java/org/apache/catalina/startup/WAS.java
@@ -1,19 +1,35 @@
 package org.apache.catalina.startup;
 
 import com.techcourse.controller.HttpController;
-import java.util.List;
-import com.techcourse.controller.Dispatcher;
+import com.techcourse.controller.ResourceFinder;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
-public class WAS implements Server {
+public class WAS {
     private final Server tomcat;
+    private final Controllers controllers;
 
-    public WAS(Server tomcat, List<HttpController> resolvers) {
+    public WAS(Server tomcat, Controllers controllers) {
         this.tomcat = tomcat;
-        resolvers.forEach(Dispatcher::register);
+        this.controllers = controllers;
     }
 
-    @Override
     public void start() {
         tomcat.start();
+    }
+
+    public static HttpResponse dispatch(HttpRequest request) throws Exception {
+        HttpResponse response = new HttpResponse();
+
+        if (controllers.contains(request.getLocation())) {
+            HttpController targetController = controllers.get(request.getLocation());
+            targetController.service(request, response);
+            return response;
+        }
+
+        String body = new ResourceFinder(request.getLocation(), request.getExtension()).getStaticResource(
+                response);
+        response.setBody(body);
+        return response;
     }
 }
